requirements.txt
discord~=2.3.2
py-cord
aiocron~=1.8
discord.py~=2.4.0
python-dotenv~=1.0.1